+++ Database/ClassifiedsDatabase/Scripts/0064-EventBookingTicket.DropColumn.Quantity.sql
@@ -1,0 +1,2 @@
+﻿ALTER TABLE EventBookingTicket
+DROP COLUMN Quantity