+++ Database/ClassifiedsDatabase/Scripts/0065-EventBookingTicket.Column.GuestDetails.sql
@@ -1,0 +1,7 @@
+﻿ALTER TABLE EventBookingTicket
+ADD GuestEmail VARCHAR(255) NULL
+
+GO
+
+ALTER TABLE EventBookingTicket
+ADD GuestFullName VARCHAR(255) NULL