+++ Paramount.Betterclassifieds.DataService/Events/EntityConfigurations/EventBookingTicketConfiguration.cs
@@ -11,8 +11,6 @@ namespace Paramount.Betterclassifieds.DataService.Events
             HasKey(prop => prop.EventBookingTicketId);
             HasRequired(prop=> prop.EventTicket).WithMany(prop => prop.EventBookingTickets).HasForeignKey(prop => prop.EventTicketId);
             HasRequired(prop => prop.EventBooking).WithMany(prop => prop.EventBookingTickets).HasForeignKey(prop => prop.EventBookingId);
-            Ignore(prop => prop.GuestFullName);
-            Ignore(prop => prop.GuestEmail);
         }
     }
 }